requirements.txt
# 核心依赖
toml>=0.10.0
peewee>=3.14.0
pydantic>=1.10.0

# gRPC依赖
grpcio>=1.50.0
grpcio-tools>=1.50.0
protobuf>=4.0.0

# RAG相关依赖
numpy>=1.20.0
sentence-transformers>=2.0.0
faiss-cpu>=1.7.0
langchain>=0.0.200

# 开发依赖
pytest>=7.0.0
black>=22.0.0
flake8>=5.0.0